requirements.txt
av
Pillow
PyYAML
tqdm
